Renaming to cnsmeth
