+++ src/cnsmeth/__about__.py
