+++ src/cnsmeth/brain_class.py
@@ -1,13 +1,13 @@
 from nicegui import ui
 
-from methnicegui.bam_handler import BamEventHandler
-from methnicegui.Sturgeon_worker import Sturgeon_worker
-from methnicegui.RCNS2_worker import RCNS2_worker
-from methnicegui.copy_number_component import CNV_Plot
-from methnicegui.target_coverage import TargetCoverage
-from methnicegui.mgmt_panel import MGMT_Panel
-from methnicegui.fusion_panel import Fusion_Panel
-from methnicegui.local_file_picker import local_file_picker
+from cnsmeth.bam_handler import BamEventHandler
+from cnsmeth.Sturgeon_worker import Sturgeon_worker
+from cnsmeth.RCNS2_worker import RCNS2_worker
+from cnsmeth.copy_number_component import CNV_Plot
+from cnsmeth.target_coverage import TargetCoverage
+from cnsmeth.mgmt_panel import MGMT_Panel
+from cnsmeth.fusion_panel import Fusion_Panel
+from cnsmeth.local_file_picker import local_file_picker
 
 from watchdog.observers import Observer
 from pathlib import Path
