+++ src/cnsmeth/theme.py
@@ -1,5 +1,5 @@
 """
-This module contains the theme for the GUI.
+This module contains the theme for the whole application.
 """
 
 from contextlib import contextmanager
@@ -7,7 +7,7 @@ from contextlib import contextmanager
 from nicegui import ui, app
 from nicegui.events import ValueChangeEventArguments
 
-from methnicegui import images
+from cnsmeth import images
 
 import os
 
@@ -19,17 +19,6 @@ IMAGEFILE = os.path.join(
 @contextmanager
 def frame(navtitle: str, myconnection):
     """Custom page frame to share the same styling and behavior across all pages"""
-    #with ui.dialog().props("persistent") as dialog, ui.card():
-    #    ui.label("Quitting the app will stop running code. Are you sure?")
-    #    ui.label("If you want to keep code running, click Cancel.")
-    #    ui.label(
-    #        "You can safely close this window and code will keep running in the background."
-    #    )
-    #    ui.button("Cancel", on_click=dialog.close).props("outline").classes("shadow-lg")
-    #    ui.button("Really Quit", icon="logout", on_click=cleanup_and_exit).props(
-    #        "outline"
-    #    ).classes("shadow-lg")
-
     with ui.header(fixed=True).classes(replace="row items-center p-2").style(
         "box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1)"
     ):
@@ -39,11 +28,7 @@ def frame(navtitle: str, myconnection):
                 ui.switch("Dark Mode", on_change=dark_mode).classes(
                     "ml-4 bg-transparent"
                 ).props('color="black"')
-                #ui.button(
-                #    "Quit", icon="logout", on_click=dialog.open
-                #)  # .classes('ml-4')#.props('outline') #.classes('shadow-lg')
                 ui.image(IMAGEFILE).style("width: 50px")
-
     with ui.column().classes("w-full"):
         yield
     with ui.footer():
