--- src/methnicegui/models/general.zip
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:706e043f3b9f248e51e9e57b0a01085efb93ae01d8a52fcff53418c7b39bbe26
-size 1777245358
