--- src/methnicegui/resources/subset_gencode.v45.annotation.gff3
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:a78e1ca0e4fe42d07c3320cb794958914e3db41b1b7cd5d9c89c25e6aca8bdf9
-size 278648447
